move tools under marius directory
