+++ marius/__init__.py
@@ -1,1 +1,4 @@
-from ._pymarius import *
+import os
+only_python = os.environ.get("MARIUS_ONLY_PYTHON", None)
+if not only_python:
+    from ._pymarius import *
