+++ marius/tools/config_templates/default_configs.txt
