+++ requirements.txt
@@ -1,0 +1,6 @@
+sphinx_rtd_theme
+breathe
+numpy
+pandas
+tox
+pytest
