+++ setup.cfg
@@ -39,7 +39,7 @@ packages =
 
 package_dir =
     marius = marius
-    marius.tools = tools
+    marius.tools = marius/tools
     marius.console_scripts = marius/console_scripts
 
 [options.entry_points]
