+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist = py
-skipsdist = true
 
 [testenv]
 commands = pytest test
-setenv = PYTHONPATH = ./tools/
+setenv =
+    MARIUS_ONLY_PYTHON = TRUE
